Did some work on the new fastqc_data files as objects idea.
I have gone back and forth on this issue, but ultimately have decided to proceed as follows:

1. Fastqc_data objects have a dictionary of modules that contain the results and raw data of their module (their name)

2. Each module will comprise its own table in the database, which will be completely flat for simplicity.

I almost decided to go with a more complicated database schema, but this is perfectly fine, since all people ever do with this data is find "how many tests failed?" kind of queries.

I also almost decided to make derived objects for each module from the module class, but realized that there wasn't that much to be done with each module -- maybe a graph or two, which is mostly provided with the original fastqc program. The raw data objects, and the dictionaries of modules the objects provide enough access for further programmatic investigation (mostly by me)