Update for Version 1.0
I have refactored most of the files for clarity and updated everything to use Python3. The working part of this script is the fastqc_results.db script, which creates a simple results database by scanning for Fastqc_results.txt files.

Here is the plan from this point (for myself mostly):
1. Update the current script to take arguments and print usage
2. Finish the select and update commands for the current version of the db
3. Add simple database viewer (HTML report or something similar)

Once these are done, I will branch and continue working on new features

3. Begin work on Full DB that stores all the information from each module, with CRUD, graphs, etc.
4. Create scripts that populate this DB and make reports from this data using simple database viewer
5. Create PyQt GUI and CLI for FastQC data, wrap in script for determining if machine has a head
6. Package and Distribute